Lots of clean up
- Moved create koin and leaderboard to commands
- Cleaned up slack.ex file
- Provide an example config for docker
