+++ .gitignore
@@ -18,4 +18,4 @@ erl_crash.dump
 .hex
 
 .env*
-config/*.env*
+config/*.env
