+++ config/docker.env.example
@@ -1,0 +1,15 @@
+HOSTNAME=localhost
+SECRET_KEY_BASE="so_secret"
+DATABASE_HOST=db
+DATABASE_PORT=5432
+DATABASE_USER=postgres
+DATABASE_PASS=postgres
+DATABASE_NAME=alex_koin_dev
+PORT=4000
+REPLACE_OS_VARS=true
+ERLANG_COOKIE=alex_koin
+MIX_ENV=dev
+SLACK_TOKEN=xoxb-THINGS_GO_HERE
+KOIN_BOT_ID=U1234
+ADMIN_ID=U5432
+KOIN_LORD_IDS=U1234,U5432
